fix: codeblock without "<" consumes extra char
Problem:
If a codeblock does not have a terminating "<" char, it consumes the
first char of the next token.

Solution:
Define (codeblock) only in terms of its lines; it doesn't need to look
for its "end". Instead, add $._line_end_codeblock to the list of things
that can terminate a (block).