feat: add `(note)` node for highlighted notes and warnings
Add support for highlighting common call-out terms in documentation,
like the legacy syntax's `helpNote`, `helpWarning`, `helpDeprecated`.

However, only the variants with a colon are supported, as these account
for 99% of use in Nvim documentation without the danger of false
positives (and to keep symbol count low).

To keep state count low, all these are parsed as a single `(note)` node
that captures can disambiguate using `#any-of?`.