revert: "ci: don't run on push"
This reverts commit d1369fcf4740e42d4ae2322d1342954f07c534c1.

The reasoning is provided by @justinmk in
https://github.com/neovim/tree-sitter-vimdoc/pull/122#issuecomment-1951443830:

It's good to save CI time, but for low-activity repos I think we should
err on the side of explicitness.

Reasons in favor of CI on push:
  - to have a baseline, so one can see the history of status checks by looking at master (including squash-merged PRs)
  - for any (rare) commits that skip the PR process

Without this, we don't have a clear concept of "current status", except
whatever the last PR was (which also assumes a linear history).